Simplify calls to setup.py
Removing the egg-info file is a bug - and it's not required by fedora
policy. There was an unclear statement on the python packaging policy
wiki that made it seem so, but it was really saying to not use upstream
eggs. egg-info dirs are important.

Additionally, it is not necessary to edit the setup.py file or to remove
the requirements files - there is an env var which turns off all code
paths related to those activities.